:recycle: refactor: move MonitoringProvider to monitoring package
